Fixed off64

+++ xen/runtime/include/sys/types.h
@@ -90,7 +90,7 @@ typedef signed long useconds_t;		/* Used for time in microseconds. */
 /* non-susv2 types: */
 __extension__ typedef signed long long loff_t;	/* 64-bit offset */
 
-__extension__ typedef signed long long off64_t;
+__extension__ typedef int64_t off64_t;
 typedef off64_t off_t;
 
 __extension__ typedef unsigned long long ino64_t;
